ab_experimentation_framework/requirements.txt
numpy==1.26.4
pandas==2.2.0
scipy==1.12.0
statsmodels==0.14.1
pyyaml==6.0.1
matplotlib==3.8.2
seaborn==0.13.2